+++ UniversityProject/Repository/IUniversityRepository.cs
@@ -4,9 +4,9 @@ using Logger;
 
 public interface IUniversityRepository
 {
-    public int Create(UniversityDto university);
-    public University Get(int ID);
+    public long Create(UniversityDto university);
+    public University Get(long ID);
     public List<University> ReturnList();
-    public void Delete(int ID);
-    public int Update(UniversityDto university);
+    public void Delete(long ID);
+    public long Update(UniversityDto university);
 }
