+++ UniversityProject/UCore/DTO/PersonalOfUniversitiesDTO.cs
@@ -2,6 +2,6 @@
 
 public class PersonalOfUniversiyDTO
 {
-    public int IdUniversity { get; set; }
+    public long IdUniversity { get; set; }
     public Administrator Administrator { get; set; }
 }
